+++ hedera-node/src/test/java/com/hedera/services/context/properties/GlobalDynamicPropertiesTest.java
@@ -97,8 +97,8 @@ class GlobalDynamicPropertiesTest {
 		assertEquals(oddCongestion, subject.congestionMultipliers());
 		assertEquals(29, subject.feesMinCongestionPeriod());
 		assertTrue(subject.autoRenewEnabled());
-		assertEquals(31L, subject.autoRenewNumberOfEntitiesToScan());
-		assertEquals(32L, subject.autoRenewMaxNumberOfEntitiesToRenewOrDelete());
+		assertEquals(31, subject.autoRenewNumberOfEntitiesToScan());
+		assertEquals(32, subject.autoRenewMaxNumberOfEntitiesToRenewOrDelete());
 		assertEquals(33L, subject.autoRenewGracePeriod());
 	}
 
@@ -142,8 +142,8 @@ class GlobalDynamicPropertiesTest {
 		assertEquals(evenCongestion, subject.congestionMultipliers());
 		assertEquals(30, subject.feesMinCongestionPeriod());
 		assertFalse(subject.autoRenewEnabled());
-		assertEquals(32L, subject.autoRenewNumberOfEntitiesToScan());
-		assertEquals(33L, subject.autoRenewMaxNumberOfEntitiesToRenewOrDelete());
+		assertEquals(32, subject.autoRenewNumberOfEntitiesToScan());
+		assertEquals(33, subject.autoRenewMaxNumberOfEntitiesToRenewOrDelete());
 		assertEquals(34L, subject.autoRenewGracePeriod());
 	}
 
@@ -183,8 +183,8 @@ class GlobalDynamicPropertiesTest {
 				.willReturn(i % 2 == 0 ? evenCongestion : oddCongestion);
 		given(properties.getIntProperty("fees.minCongestionPeriod")).willReturn(i + 28);
 		given(properties.getBooleanProperty("autorenew.isEnabled")).willReturn((i + 29) % 2 == 0);
-		given(properties.getLongProperty("autorenew.numberOfEntitiesToScan")).willReturn(i + 30L);
-		given(properties.getLongProperty("autorenew.maxNumberOfEntitiesToRenewOrDelete")).willReturn(i + 31L);
+		given(properties.getIntProperty("autorenew.numberOfEntitiesToScan")).willReturn(i + 30);
+		given(properties.getIntProperty("autorenew.maxNumberOfEntitiesToRenewOrDelete")).willReturn(i + 31);
 		given(properties.getLongProperty("autorenew.gracePeriod")).willReturn(i + 32L);
 	}
